+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 2021.03.14-23.22.46
+
 - Fix completion inside refers, re-add support for it - Fixes #364
 - Change range of expression functions clj-kondo diagnostics to avoid collision with function signature.